requirements.txt
# packages in environment at /home/kevin/anaconda3/envs/tf-gpu:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
_tflow_select             2.1.0                       gpu  
absl-py                   0.9.0                    py36_0  
astor                     0.8.0                    py36_0  
astroid                   2.4.1                    py36_0  
astunparse                1.6.3                      py_0  
backports                 1.0                        py_2  
backports.weakref         1.0rc1                   py36_0  
blas                      1.0                         mkl  
bleach                    1.5.0                    py36_0  
blinker                   1.4                      py36_0  
c-ares                    1.15.0            h7b6447c_1001  
ca-certificates           2020.1.1                      0  
cachetools                3.1.1                      py_0  
certifi                   2020.4.5.1               py36_0  
cffi                      1.14.0           py36he30daa8_1  
chardet                   3.0.4                 py36_1003  
click                     7.1.2                      py_0  
cryptography              2.9.2            py36h1ba5d50_0  
cudatoolkit               10.0.130                      0  
cudnn                     7.6.5                cuda10.0_0  
cupti                     10.0.130                      0  
gast                      0.3.3                      py_0  
google-auth               1.14.1                     py_0  
google-auth-oauthlib      0.4.1                      py_2  
google-pasta              0.2.0                      py_0  
grpcio                    1.27.2           py36hf8bcb03_0  
h5py                      2.10.0           py36h7918eee_0  
hdf5                      1.10.4               hb1b8bf9_0  
html5lib                  0.9999999                py36_0  
idna                      2.9                        py_1  
intel-openmp              2020.1                      217  
isort                     4.3.21                   py36_0  
keras                     2.3.1                         0  
keras-applications        1.0.8                      py_0  
keras-base                2.3.1                    py36_0  
keras-preprocessing       1.1.0                      py_1  
lazy-object-proxy         1.4.3            py36h7b6447c_0  
ld_impl_linux-64          2.33.1               h53a641e_7  
libedit                   3.1.20181209         hc058e9b_0  
libffi                    3.3                  he6710b0_1  
libgcc-ng                 9.1.0                hdf63c60_0  
libgfortran-ng            7.3.0                hdf63c60_0  
libprotobuf               3.11.4               hd408876_0  
libstdcxx-ng              9.1.0                hdf63c60_0  
markdown                  3.1.1                    py36_0  
mccabe                    0.6.1                    py36_1  
mkl                       2020.1                      217  
mkl-service               2.3.0            py36he904b0f_0  
mkl_fft                   1.0.15           py36ha843d7b_0  
mkl_random                1.1.1            py36h0573a6f_0  
mock                      4.0.2                      py_0  
ncurses                   6.2                  he6710b0_1  
numpy                     1.18.1           py36h4f9e942_0  
numpy-base                1.18.1           py36hde5b4d6_1  
oauthlib                  3.1.0                      py_0  
opencv-python             4.2.0.34                 pypi_0    pypi
openssl                   1.1.1g               h7b6447c_0  
opt_einsum                3.1.0                      py_0  
pillow                    7.1.2                    pypi_0    pypi
pip                       20.0.2                   py36_3  
protobuf                  3.11.4           py36he6710b0_0  
pyasn1                    0.4.8                      py_0  
pyasn1-modules            0.2.7                      py_0  
pycparser                 2.20                       py_0  
pyjwt                     1.7.1                    py36_0  
pylint                    2.5.2                    py36_0  
pyopenssl                 19.1.0                   py36_0  
pysocks                   1.7.1                    py36_0  
python                    3.6.10               h7579374_2  
pyyaml                    5.3.1            py36h7b6447c_0  
readline                  8.0                  h7b6447c_0  
requests                  2.23.0                   py36_0  
requests-oauthlib         1.3.0                      py_0  
rsa                       4.0                        py_0  
scipy                     1.4.1            py36h0b6359f_0  
setuptools                46.4.0                   py36_0  
six                       1.14.0                   py36_0  
sqlite                    3.31.1               h62c20be_1  
tensorboard               1.13.1           py36hf484d3e_0  
tensorboard-plugin-wit    1.6.0                      py_0  
tensorflow                1.13.1          gpu_py36h3991807_0  
tensorflow-base           1.13.1          gpu_py36h8d69cac_0  
tensorflow-estimator      1.13.0                     py_0  
tensorflow-gpu            1.13.1               h0d30ee6_0  
termcolor                 1.1.0                    py36_1  
tk                        8.6.8                hbc83047_0  
toml                      0.10.0             pyh91ea838_0  
typed-ast                 1.4.1            py36h7b6447c_0  
urllib3                   1.25.8                   py36_0  
werkzeug                  1.0.1                      py_0  
wheel                     0.34.2                   py36_0  
wrapt                     1.11.2           py36h7b6447c_0  
xz                        5.2.5                h7b6447c_0  
yaml                      0.1.7                had09818_2  
zlib                      1.2.11               h7b6447c_3  
